+++ CHANGELOG.md
@@ -11,6 +11,7 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 ### Added
 
 * Add CITATIONS.cff
+* Developer's Common Tasks
 
 ### Changed